All tests and implementations are now complete

+++ istring_rep.h
@@ -8,7 +8,7 @@ typedef struct _istring {
 } _istring, *Istring;
 
 // Macros
-#define START(p) ((Istring) (p - sizeof(int)))
-#define STRING(p) ((char*) (p + sizeof(int)))
+#define START(p) ((Istring) (p - sizeof(size_t)))
+#define STRING(p) ((char*) (p + sizeof(size_t)))
 
 #endif
